refactor(tests): testhelpers pkg
